feat: working "koe-kak"

+++ src/adc.h
@@ -13,5 +13,5 @@ void start_conversion(void);
  * Функция получения очередного значения АЦП
  * @returns значение полученное с АЦП в диапазоне от 0-1
  */
-double read_adc(void);
+uint16_t read_adc(void);
 #endif
